streamlit-dashboard/requirements.txt
streamlit
requests
python-dotenv
pydantic-settings
pandas
plotly